chore: rename ThemeTrans to Trans
