+++ packages/gatsby-theme-newrelic/src/components/PagingInfoView.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import ThemeTrans from './ThemeTrans';
+import Trans from './Trans';
 
 import { appendClassName } from '@elastic/react-search-ui-views/es/view-helpers';
 
@@ -15,19 +15,16 @@ function PagingInfo({
   ...rest
 }) {
   return (
-    <ThemeTrans
-      i18nKey="search.pagingInfo"
-      parent="div"
-      className={appendClassName('sui-paging-info', className)}
-      {...rest}
-    >
-      Showing{' '}
-      <strong>
-        {start} - {end}
-      </strong>{' '}
-      out of <strong>{totalResults}</strong> for <em>{searchTerm}</em> across{' '}
-      {sites.length} sites
-    </ThemeTrans>
+    <div className={appendClassName('sui-paging-info', className)} {...rest}>
+      <Trans i18nKey="search.pagingInfo">
+        Showing{' '}
+        <strong>
+          {start} - {end}
+        </strong>{' '}
+        out of <strong>{totalResults}</strong> for <em>{searchTerm}</em> across{' '}
+        {sites.length} sites
+      </Trans>
+    </div>
   );
 }
